requirements.txt
# CGRN — Core Requirements
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
scikit-learn>=1.2.0
numpy>=1.23.0
tqdm>=4.64.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Optional: FLOPs computation
thop>=0.1.1

# Optional: Progress bars in training
tqdm>=4.64.0
